+++ src/vertex/core/urls.tsx
@@ -15,7 +15,7 @@ export const DEVICES_MGT_URL = `${BASE_API_URL}/devices`;
 export const SITES_MGT_URL = `${BASE_API_URL}/devices/sites`;
 export const ANALYTICS_MGT_URL = `${BASE_API_URL}/analytics`;
 
-const ANALYTICS_BASE_URL = stripTrailingSlash(
+export const ANALYTICS_BASE_URL = stripTrailingSlash(
   process.env.NEXT_PUBLIC_ANALYTICS_URL || DEFAULT_ANALYTICS_BASE_URL
 );
 export const forgotPasswordUrl = `${ANALYTICS_BASE_URL}/user/forgotPwd`;